Factorio: client cleanup and prevent process bomb
