+++ worlds/factorio/__init__.py
@@ -5,7 +5,6 @@ import logging
 import typing
 
 import Utils
-import settings
 from BaseClasses import Region, Location, Item, Tutorial, ItemClassification
 from worlds.AutoWorld import World, WebWorld
 from worlds.LauncherComponents import Component, components, Type, launch as launch_component
@@ -20,6 +19,7 @@ from .Technologies import base_tech_table, recipe_sources, base_technology_table
     progressive_technology_table, common_tech_table, tech_to_progressive_lookup, progressive_tech_table, \
     get_science_pack_pools, Recipe, recipes, technology_table, tech_table, factorio_base_id, useless_technologies, \
     fluids, stacking_items, valid_ingredients, progressive_rows
+from .settings import FactorioSettings
 
 
 def launch_client():
@@ -28,29 +28,6 @@ def launch_client():
 
 
 components.append(Component("Factorio Client", "FactorioClient", func=launch_client, component_type=Type.CLIENT))
-
-
-class FactorioSettings(settings.Group):
-    class Executable(settings.UserFilePath):
-        is_exe = True
-
-    class ServerSettings(settings.OptionalUserFilePath):
-        """
-        by default, no settings are loaded if this file does not exist. \
-If this file does exist, then it will be used.
-        server_settings: "factorio\\\\data\\\\server-settings.json"
-        """
-
-    class FilterItemSends(settings.Bool):
-        """Whether to filter item send messages displayed in-game to only those that involve you."""
-
-    class BridgeChatOut(settings.Bool):
-        """Whether to send chat messages from players on the Factorio server to Archipelago."""
-
-    executable: Executable = Executable("factorio/bin/x64/factorio")
-    server_settings: typing.Optional[FactorioSettings.ServerSettings] = None
-    filter_item_sends: typing.Union[FilterItemSends, bool] = False
-    bridge_chat_out: typing.Union[BridgeChatOut, bool] = True
 
 
 class FactorioWeb(WebWorld):
